fix: avg_rating toFixed 크래시 — PostgreSQL numeric→float 캐스팅
pg 라이브러리가 numeric 타입을 string으로 반환 → "4.2".toFixed() TypeError.
SQL에서 ::float 캐스팅, 컴포넌트에서 Number() 방어 처리 추가.

+++ web/app/api/coupang/stats/route.ts
@@ -15,7 +15,7 @@ export async function GET() {
       SELECT
         (SELECT COUNT(*) FROM reviews)::int           AS total_reviews,
         (SELECT COUNT(*) FROM products)::int          AS total_products,
-        ROUND(AVG(rating)::numeric, 2)                AS avg_rating,
+        ROUND(AVG(rating)::numeric, 2)::float         AS avg_rating,
         MAX(collected_at)                             AS last_updated
       FROM reviews
     `)
